refactor(pi-codex-conversion): isolate exec process lifecycle
